Added QA mode: only display debug information to specified IPs

+++ XoopsEngine/htdocs/clear.php
@@ -27,7 +27,7 @@ $path = XOOPS::path("var") . "/cache/";
 $objects = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($path), RecursiveIteratorIterator::SELF_FIRST);
 $count = 0;
 foreach ($objects as $object) {
-    if ($object->isFile()) {
+    if ($object->isFile() && 'index.html' !== $object->getFilename()) {
         unlink($object->getPathname());
         $count++;
     }
